Hack to avoid sequences of three Tx ops followed by Rx.
This seems to break the last Rx for whatever reason.  I still need to
test this with more MCUs and more nRFs, but it looks like sending exactly
three Tx packets in a row cause the subsequent Rx packet to not be
reported on the SPI bus.  If there are fewer or more Tx operations,
everything is fine, only 3 TXs in a row cause this problem regardless of
the timing and the sizes of the packets.  So we try to count subsequent
Tx operations and split the 3rd one in a row into two packets.

Oh, well.