Change default CSN pin assignment in favot of SS.
Since there's only one general use pin left in the continuous group of
pins on this side of many Atmega328 based boards, make sure the pin that
is left is capable of servicing the nRF24L01+'s Interrupt pin.  SS is
not capable -- some state transitions on it will disable SPI master mode
and kill communication.